+++ docs/router/index.js
@@ -17,6 +17,7 @@ function route(path, component) {
 
 // Templates
 const Documentation = () => import('@/templates/Documentation')
+const NotFound = () => import('@/pages/404')
 
 const router = new Router({
     mode: 'history',
@@ -80,7 +81,14 @@ const router = new Router({
             ]
         },
         // Global redirect
-        { path: '*', redirect: '' }
+        // { path: '*', redirect: '404' }
+        // 404 page
+        {
+            path: '*',
+            name: 'notfound',
+            meta: routes['notfound'],
+            component: NotFound
+        }
     ],
     scrollBehavior(to, from, savedPosition) {
         if (to.hash) {