+++ src/components/navbar/NavbarDropdown.spec.js
@@ -16,4 +16,20 @@ describe('BNavbarDropdown', () => {
     it('render correctly', () => {
         expect(wrapper.html()).toMatchSnapshot()
     })
+
+    it('sets newActive when active prop is changed', () => {
+        wrapper.setProps({active: true})
+        expect(wrapper.vm.newActive).toBeTruthy()
+
+        wrapper.setProps({active: false})
+        expect(wrapper.vm.newActive).toBeFalsy()
+    })
+
+    it('sets newActive when menu is shown/closed', () => {
+        wrapper.vm.showMenu()
+        expect(wrapper.vm.newActive).toBeTruthy()
+
+        wrapper.vm.closeMenu()
+        expect(wrapper.vm.newActive).toBeFalsy()
+    })
 })