+++ src/components/radio/RadioButton.vue
@@ -27,50 +27,21 @@
 </template>
 
 <script>
+import CheckRadioMixin from '../../utils/CheckRadioMixin.js'
+
 export default {
     name: 'BRadioButton',
+    mixins: [CheckRadioMixin],
     props: {
-        value: [String, Number, Boolean, Function, Object, Array],
-        nativeValue: [String, Number, Boolean, Function, Object, Array],
         type: {
             type: String,
             default: 'is-primary'
         },
-        disabled: Boolean,
-        required: Boolean,
-        expanded: Boolean,
-        name: String,
-        size: String
+        expanded: Boolean
     },
     data() {
         return {
-            newValue: this.value,
             isFocused: false
-        }
-    },
-    computed: {
-        computedValue: {
-            get() {
-                return this.newValue
-            },
-            set(value) {
-                this.newValue = value
-                this.$emit('input', value)
-            }
-        }
-    },
-    watch: {
-        /**
-        * When v-model change, set internal value.
-        */
-        value(value) {
-            this.newValue = value
-        }
-    },
-    methods: {
-        focus() {
-            // MacOS FireFox and Safari do not focus when clicked
-            this.$refs.input.focus()
         }
     }
 }