+++ src/components/snackbar/index.js
@@ -2,24 +2,27 @@ import Vue from 'vue'
 import Snackbar from './Snackbar'
 
 import config from '../../utils/config'
+import { merge } from '../../utils/helpers'
 import { use, registerComponentProgrammatic } from '../../utils/plugins'
 
 const SnackbarProgrammatic = {
     open(params) {
-        let message
         let parent
-        if (typeof params === 'string') message = params
+        if (typeof params === 'string') {
+            params = {
+                message: params
+            }
+        }
 
         const defaultParam = {
             type: 'is-success',
-            position: config.defaultSnackbarPosition || 'is-bottom-right',
-            message
+            position: config.defaultSnackbarPosition || 'is-bottom-right'
         }
         if (params.parent) {
             parent = params.parent
             delete params.parent
         }
-        const propsData = Object.assign(defaultParam, params)
+        const propsData = merge(defaultParam, params)
 
         const vm = typeof window !== 'undefined' && window.Vue ? window.Vue : Vue
         const SnackbarComponent = vm.extend(Snackbar)
@@ -42,5 +45,6 @@ use(Plugin)
 export default Plugin
 
 export {
-    SnackbarProgrammatic
+    SnackbarProgrammatic,
+    Snackbar as BSnackbar
 }