+++ src/index.js
@@ -1,12 +1,13 @@
 import * as components from './components'
 
+import { merge } from './utils/helpers'
 import config, { setOptions } from './utils/config'
 import { use, registerComponentProgrammatic } from './utils/plugins'
 
 const Buefy = {
     install(Vue, options = {}) {
         // Options
-        setOptions(Object.assign(config, options))
+        setOptions(merge(config, options, true))
         // Components
         for (let componentKey in components) {
             Vue.use(components[componentKey])
@@ -14,7 +15,7 @@ const Buefy = {
         // Config component
         const BuefyProgrammatic = {
             setOptions(options) {
-                setOptions(Object.assign(config, options))
+                setOptions(merge(config, options, true))
             }
         }
         registerComponentProgrammatic(Vue, 'config', BuefyProgrammatic)