+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Buefy Changelog
 
+## 0.8.7
+
+### New features
+
+* Add ``controls`` prop to number input (thanks @afwn90cj93201nixr2e1re)
+* Add ``accordion`` prop to menu
+* Add ``icon-click`` event to input (thanks @service-paradis)
+* Add multiple support to Datepicker when type is ``month`` 
+* Add more customization variables for dropdown and loading (thanks @HugoHeneault)
+* Add more customization variables for datepicker (thanks @service-paradis)
+
+### Fixes
+
+* Fix IE compatibility using multiple datepicker
+* Fix datetimepicker ignores expanded prop (thanks @cypressious)
+* Fix disabled checkbox in table (thanks @yxngl)
+* Fix page number count when filtering data (thanks @yxngl)
+* Fix placeholder to datetimepicker (thanks @yxngl)
+* Fix #1812 extending a component
+* Fix #1909, #1902, #1898 issues on searchable table column (thanks @dauriata and @service-paradis )
+* Fix #1951 modal a11y improvements (thanks @service-paradis)
+* Fix #1963 erase number-input value
+* Fix #1972 using min and max native datepicker and timepicker props
+* Fix #1986 setting 00:00 on datetimepicker
+
+
 ## 0.8.6
 
 ### New features